fixed signed/unsigned comparison error

+++ portage/distributed/mpi_bounding_boxes.h
@@ -92,7 +92,7 @@ class MPI_Bounding_Boxes {
     MPI_Comm_rank(comm_, &commRank);
 
     dim_ = source_mesh.space_dimension();
-    assert(dim_ == target_mesh.space_dimension());
+    assert(dim_ == static_cast<int>(target_mesh.space_dimension()));
 
     // sendFlags, which partitions to send data
     // this is computed via intersection of whole partition bounding boxes
